Mix reverse translations with forward ones in the quiz
Forward translations only help with understanding the language. Reverse
translations will prompt us with the name in our native language, and we
will have to remember the foreign one. This helps speaking and writing as
opposed to listening and reading.

As I was touching this bit of code, I have changed a variable to camelCase
instead of underscored, to be consistent with the rest of the code.