Make the "translation" field in the Word non-nullable
It makes sense for every word in the DB to have a translation, i.e. a
new word learnt must mean something in your native language.

Even though the code is trivial, it changes the DB schema. Therefore, the
DB version number goes up to 2, which means 1) a DB migration path must be
provided for app updates from DB v.1 to DB v.2 as trivial as it can be, and
2) the migration path must be tested. This is the first time we increase
the DB version, and the first time we add testing for the app (yay!), so I
am adding a lot of new boilerplate code in this commit to enable both
migrations and testing.