Implement asynchronous DB updates
By using LiveData and the observer pattern, activities now observe the
dictionary and get notified every time it changes. This is done
asynchronously in order to avoid blocking the app, as it happened before
with large dictionaries every time we queried the dictionary (since it had
to query the whole word list every time).